Revert "charger: Draw battery percentage over charger screen"
This reverts commit f7e1ef2461121c1ebe7f1ef3ea25cca397e4c5ab.

Rationale:
The tiny (blue) number drawn on top of the otherwise black-and-white
battery symbol looks pretty out of place and gives the perception of
being a debugging leftover that just was forgotten to be removed.
Additionally it's very unlikely the user is interested in the exact
value of battery charge in offmode charging. It's much more likely he's
interested in rough state information only ('is my battery fully charged
now?').

Change-Id: I2b4f881faa4a254d03e07d5d46b6870a2fe9ec4a